requirements.txt
anyio=3.6.1
certifi=2022.6.15.1
coloredlogs=15.0.1
ddddocr=1.4.7
flatbuffers=2.0.7
h11=0.12.0
httpcore=0.15.0
httpx=0.23.0
humanfriendly=10.0
idna=3.3
jiphy=1.2.2
lxml=4.9.1
mpmath=1.2.1
numpy=1.23.3
onnxruntime=1.12.1
opencv-python-headless=4.6.0.66
packaging=21.3
Pillow=9.2.0
pip=22.2.1
protobuf=4.21.5
pycryptodome=3.15.0
pyDes=2.0.1
PyExecJS=1.5.1
pyparsing=3.0.9
pyreadline3=3.4.1
rfc3986=1.5.0
setuptools=63.2.0
six=1.16.0
sniffio=1.3.0
sympy=1.11.1